+++ bin/install.dart
@@ -1,0 +1,5 @@
+import 'package:easy_notify/src/installer.dart';
+
+void main() async {
+  await EasyNotifyInstaller.run();
+}