+++ lib/easy_notify.dart
@@ -1,3 +1,5 @@
 library easy_notify;
 
 export 'src/easy_notify_core.dart';
+export 'src/easy_notify_permissions.dart';
+export 'src/installer.dart';